Move to using read/write item caches.
