+++ src/options.rs
@@ -29,7 +29,8 @@ pub enum OpenMode {
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub struct OpenOptions {
     pub strands: Option<u32>,
-    pub cache: Option<usize>,
+    pub read_cache: Option<usize>,
+    pub write_cache: Option<usize>,
     pub mode: OpenMode,
 }
 
@@ -37,7 +38,8 @@ impl OpenOptions {
     pub fn new() -> Self {
         OpenOptions {
             strands: None,
-            cache: None,
+            read_cache: None,
+            write_cache: None,
             mode: OpenMode::Open,
         }
     }
@@ -47,8 +49,13 @@ impl OpenOptions {
         self
     }
 
-    pub fn cache(&mut self, bytes: usize) -> &mut Self {
-        self.cache = Some(bytes);
+    pub fn read_cache(&mut self, bytes: usize) -> &mut Self {
+        self.read_cache = Some(bytes);
+        self
+    }
+
+    pub fn write_cache(&mut self, bytes: usize) -> &mut Self {
+        self.write_cache = Some(bytes);
         self
     }
